requirements.txt
Django
djangorestframework
mysqlclient
python-dateutil